Refactoring: Better handling of gint64 type

+++ src/python/contentstat-py.c
@@ -119,7 +119,7 @@ get_num(_ContentStatObject *self, void *member_offset)
     if (check_ContentStatStatus(self))
         return NULL;
     cr_ContentStat *rec = self->stat;
-    gint64 val = *((gint64 *) ((size_t)rec + (size_t) member_offset));
+    gint64 val = (gint64) *((gint64 *) ((size_t)rec + (size_t) member_offset));
     return PyLong_FromLongLong((long long) val);
 }
 
@@ -129,7 +129,7 @@ get_int(_ContentStatObject *self, void *member_offset)
     if (check_ContentStatStatus(self))
         return NULL;
     cr_ContentStat *rec = self->stat;
-    gint64 val = *((int *) ((size_t)rec + (size_t) member_offset));
+    gint64 val = (gint64) *((int *) ((size_t)rec + (size_t) member_offset));
     return PyLong_FromLongLong((long long) val);
 }
